% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{charger_packages}
\alias{charger_packages}
\title{Charger les packages nécessaires}
\usage{
charger_packages()
}
\description{
Cette fonction charge les packages nécessaires pour le package olympicsWeather.
}
