% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weather_forecast.R
\name{forecast.character}
\alias{forecast.character}
\title{Prévisions météorologiques à partir d'une adresse.}
\usage{
forecast.character(adresse)
}
\arguments{
\item{adresse}{Adresse à géocoder}
}
\value{
Les prévisions météorologiques obtenues
}
\description{
Prévisions météorologiques à partir d'une adresse.
}
