% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weather_forecast.R
\name{forecast.numeric}
\alias{forecast.numeric}
\title{Verification du type "numeric" du parametre.}
\usage{
forecast.numeric(coordonnée)
}
\arguments{
\item{coordonnée}{Vecteur numerique de taille 2.}
}
\value{
"L'argument coordonnee doit etre un vecteur numerique de taille 2 (latitude, longitude)." Si le vecteur ne respecte pas le bon typage.
}
\description{
Verification du type "numeric" du parametre.
}
