% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weather_forecast.R
\name{get_forecast.character}
\alias{get_forecast.character}
\title{get_forecast.character}
\usage{
\method{get_forecast}{character}(address)
}
\arguments{
\item{address}{a character, an address}
}
\value{
une liste comprenant les données sous forme de graphiques, puis de tableau de 5 colonnes
}
\description{
Cette fonction permet d'obtenir des prévisions météo facilement lisibles sur
1 semaine, heure par heure, à partir du nom ou de l'adresse d'un lieu. Elles
sont issues de l'API Open Météo.
}
\examples{
get_forecast.character("Parc des Princes")
}
