% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_forecast_numeric.R
\name{get_forecast.numeric}
\alias{get_forecast.numeric}
\title{get_forecast.numeric}
\usage{
get_forecast.numeric(xy)
}
\arguments{
\item{xy}{un vecteur numerique de taille 2}
}
\value{
un tibble des previsions meteorologiques
}
\description{
Cette fonction consiste a donner les previsions meteorologiques a partir des coordonnees fournies
}
\examples{
xy <- c(47.7,-3.37)
previsions <- get_forecast.numeric(xy)
head(previsions)
}
