% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weather_forecast.R
\name{get_forecast}
\alias{get_forecast}
\title{Get Weather Forecast}
\usage{
get_forecast(add_ou_coord)
}
\arguments{
\item{add_ou_coord}{A numeric vector of length 2 representing coordinates (latitude and longitude) or a character string representing an address.}
}
\value{
A tibble containing the weather forecast.
}
\description{
This function retrieves the weather forecast based on either coordinates (latitude and longitude) or an address.
}
\examples{
get_forecast(c(48.85, 2.35)) # Using coordinates
get_forecast("1600 Amphitheatre Parkway, Mountain View, CA") # Using address

}
