% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weather_forecast.R
\name{get_forecast_visualisation}
\alias{get_forecast_visualisation}
\title{get_forecast_visualisation}
\usage{
get_forecast_visualisation(location)
}
\arguments{
\item{location(coordonnées_géographiques_ou_nom_du_site)}{}
}
\value{
quatre graphiques intéractifs
}
\description{
get_forecast_visualisation
}
