% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weather_forecast.R
\name{get_gps_coordinate}
\alias{get_gps_coordinate}
\title{Récupère les coordonnées GPS pour une adresse donnée.}
\usage{
get_gps_coordinate(address)
}
\arguments{
\item{address}{Une adresse sous forme de texte.}
}
\value{
Un vecteur numérique de taille 2 avec les coordonnées GPS (latitude, longitude).
}
\description{
Récupère les coordonnées GPS pour une adresse donnée.
}
