% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weather_forecast.R
\name{get_graphs}
\alias{get_graphs}
\title{get_graphs}
\usage{
get_graphs(data)
}
\arguments{
\item{data}{a tibble with 5 columns}
}
\value{
4 plots, the evolution of the values
}
\description{
Cette fonction affiche 4 graphiques représentant chacun l'évolution d'une des
caractéristiques météo prévues.
}
\examples{
get_graphs(get_forecast(c(48.841363,2.253069)))

}
