% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weather_forecast.R
\name{get_weather_forecast}
\alias{get_weather_forecast}
\title{get_weather_forecast}
\usage{
get_weather_forecast(arg)
}
\arguments{
\item{arg}{Soit un vecteur numerique de taille 2 : c(x,y), soit une adresse textuelle}
}
\value{
un tibble sous format JSON
}
\description{
Cette fonction renvoie les previsions meteorologiques selon l'adresse ou les coordonnees fournies.
Nous avons des informations sur la temperature ressentie et celle attendue, des chances de pluie et de la quantite de pluie observee.
Si l'argument est un vecteur, la fonction fait appel a la fonction get_forecast.numeric.
Et si c'est une adresse, elle fait appel a la fonction get_forecast.character.
Sinon, elle ne renvoie rien si l'argument ne correspond a aucun des deux types.
}
\examples{
get_weather_forecast(c(48.85,2.35))
get_weather_forecast("Paris, France")
}
