% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weather_forecast.R
\name{graph_function}
\alias{graph_function}
\title{Fonction pour tracer les prévisions météo.}
\usage{
graph_function(result)
}
\arguments{
\item{result}{Une tibble contenant les prévisions météo.}
}
\value{
Un graphique des prévisions météo.
}
\description{
Cette fonction prend en entrée les prévisions météo sous forme de tibble et trace un graphique.
}
