% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualisation.R
\name{Graphique_pluie}
\alias{Graphique_pluie}
\title{Graphique représentant la précipitation en mm et la probabilité de pluie en \%}
\usage{
Graphique_pluie(resultats)
}
\arguments{
\item{resultats}{Un tibble retourné par la fonction get_forecast}
}
\value{
Un graphique de visualisation de la précipitation en mm et de la probabilité de pluie à un point/adresse données
}
\description{
Graphique représentant la précipitation en mm et la probabilité de pluie en \%
}
