% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualisation.R
\name{Graphique_temperature}
\alias{Graphique_temperature}
\title{Graphique Représentant la température et la température en degrés ccelsius}
\usage{
Graphique_temperature(resultats)
}
\arguments{
\item{resultats}{Un tibble retourné par la fonction get_forecast}
}
\value{
Un graphique de visualisation de la température et la température en degrés ccelsius à un point/adresse données
}
\description{
Graphique Représentant la température et la température en degrés ccelsius
}
