% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olympicsWeather-package.R
\docType{package}
\name{olympicsWeather-package}
\alias{olympicsWeather}
\alias{olympicsWeather-package}
\title{olympicsWeather: olympicsWeather}
\description{
Une librarie R pour obtenir des prévisions météorologiqu et visualiser les données associées.
}
\author{
\strong{Maintainer}: Rodolphe Mocaer \email{Rodolphe.mocaer@gmail.com}

}
\keyword{internal}
