% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weather_forecast.R
\name{perform_request}
\alias{perform_request}
\title{perform_request}
\usage{
perform_request(lat, lon)
}
\arguments{
\item{lat}{numeric Latitude coordinate.}

\item{lon}{numeric Longitude coordinate.}
}
\value{
A tibble with 5 rows containing weather forecast information.
}
\description{
This function is designed to create a tibble by making a request to an API.
}
\details{
This function takes latitude and longitude coordinates as input, constructs a URL to query a weather forecast API, and retrieves weather forecast data in the form of a tibble with 5 rows.

The URL used for the API request is "https://api.open-meteo.com/v1/forecast". The function then sends a request to this URL with specified query parameters, including latitude, longitude, and requested hourly weather forecast data such as temperature, apparent temperature, precipitation probability, and precipitation amount.

The function returns the retrieved weather forecast data as a tibble.
}
\examples{
perform_request(48.85, 2.35)

}
\seealso{
\code{\link[=request]{request()}}, \code{\link[=req_url_query]{req_url_query()}}, \code{\link[=req_perform]{req_perform()}}, \code{\link[=resp_body_json]{resp_body_json()}}, \code{\link[=as_tibble]{as_tibble()}}
}
