% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weather_forecast.R
\name{unnest_data}
\alias{unnest_data}
\title{Transformer les Données Météo en Format Tibble}
\usage{
unnest_data(meteo_data)
}
\arguments{
\item{meteo_data}{Un tibble contenant les données brutes de l'API météorologique.}
}
\value{
Un tibble structuré avec les colonnes pour la date, l'heure, la température,
la température ressentie, la probabilité de précipitation et les précipitations.
}
\description{
Cette fonction prend un ensemble de données météorologiques horaires et les transforme
en un tibble plus lisible, en extrayant des informations spécifiques comme la date, l'heure,
la température, et la probabilité de précipitation.
}
\examples{
perform_request(lat = 47.21725, long = -1.55336) |> unnest_data()
}
