% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weather_forecast.R
\name{unnest_response}
\alias{unnest_response}
\title{unnest_response}
\usage{
unnest_response(resp)
}
\arguments{
\item{resp}{Weather API response, typically in the form of a list.}
}
\value{
A tibble with columns for date and time, temperature in Celsius, apparent temperature in Celsius, precipitation probability, and precipitation amount.
}
\description{
This function is designed to flatten weather API response data into a tibble.
}
\details{
This function takes as input the weather API response, typically in the form of a list, and extracts relevant data to flatten into a tibble with the following columns:
\itemize{
\item date_time: Date and time of the weather forecast.
\item temperature_celsius: Temperature in Celsius.
\item apparent_temperature_celsius: Apparent temperature in Celsius.
\item precipation_probability: Probability of precipitation.
\item precipitation: Amount of precipitation.
}

The function then returns this flattened tibble with the data.
}
\examples{
unnest_response(response_table)

}
\seealso{
\code{\link[=as_tibble]{as_tibble()}}
}
