% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weather_forecast.R
\name{visualiser_carte}
\alias{visualiser_carte}
\title{Visualise les emplacements des prévisions météorologiques sur une carte Leaflet}
\usage{
visualiser_carte(forecast_locations)
}
\arguments{
\item{forecast_locations}{Un data frame contenant les emplacements des prévisions météorologiques avec au moins deux colonnes : long (longitude) et lat (latitude).}
}
\value{
Une carte Leaflet avec des marqueurs pour les emplacements des prévisions météorologiques.
}
\description{
Visualise les emplacements des prévisions météorologiques sur une carte Leaflet
}
