% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weather_forecast.R
\name{visualiser_forecast_h}
\alias{visualiser_forecast_h}
\title{Visualisation METEO heure}
\usage{
visualiser_forecast_h(x, y = NULL)
}
\arguments{
\item{x}{Latitude (num) ou adresse (string).}

\item{y}{Longitude (num) (Obligatoire uniquement si x est la latitude).}
}
\value{
Des graphiques ggplot
}
\description{
Cette fonction permet de visualiser les donnees de previsions meteorologiques par heures en fonction de la latitude et de la longitude ou d une adresse.
}
\examples{
visualiser_forecast_h(42., 2.65)
visualiser_forecast_h("Nantes")
}
