% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weather_forecast.R
\name{visualiser_previsions}
\alias{visualiser_previsions}
\title{Visualise les prévisions météorologiques}
\usage{
visualiser_previsions(forecast_data)
}
\arguments{
\item{forecast_data}{Un data frame contenant les données de prévisions météorologiques avec au moins deux colonnes : date et température.}
}
\value{
Un objet ggplot représentant les prévisions météorologiques.
}
\description{
Visualise les prévisions météorologiques
}
