% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weather_forecast.R
\name{visualiser_temperatures}
\alias{visualiser_temperatures}
\title{Visualiser les Températures avec Interactivité Plotly}
\usage{
visualiser_temperatures(weather_forecast)
}
\arguments{
\item{weather_forecast}{Tibble contenant les données météorologiques à visualiser.
Le Tibble ou data frame doit contenir au moins les colonnes \code{date_heure} pour les dates et heures des prévisions,
\code{temperature_celsius} pour les températures en degrés Celsius, et \code{temperature_ressentie}
pour les températures ressenties, également en degrés Celsius.}
}
\value{
Un graphique interactif \code{plotly} est affiché.
}
\description{
Cette fonction génère un graphique interactif affichant la température et la température ressentie
au fil du temps. Elle utilise \code{ggplot2} pour créer le graphique initial puis le convertit
en un graphique interactif avec \code{plotly}.
}
\examples{
# Supposer que weather_data est un data frame contenant
# les colonnes date_heure, temperature_celsius et temperature_ressentie
# weather_data <- data.frame(
#   date_heure = seq(as.POSIXct("2022-01-01"), by = "hour", length.out = 24),
#   temperature_celsius = rnorm(24, mean = 15, sd = 5),
#   temperature_ressentie = rnorm(24, mean = 15, sd = 5)
# )
# visualiser_temperatures(weather_data)

}
