% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_weather_forecast.R
\name{visualize_weather_forecast}
\alias{visualize_weather_forecast}
\title{Visualise les prévisions météorologiques de manière graphique.}
\usage{
visualize_weather_forecast(localisation)
}
\arguments{
\item{localisation}{Latitude et longitude de la coordonnée GPS ou d'une adresse .}

\item{forecast}{Résultats obtenus à partir de la fonction \code{get__weather_forecast}.}
}
\value{
Un graphique représentant les prévisions météorologiques.
}
\description{
Cette fonction prend en entrée les résultats obtenus à partir de la fonction \code{get__weather_forecast} et produit une sortie visuelle, telle qu'un graphique représentant les températures prévues.
}
\examples{
visualize_weather_forecast("Parc des princes")
visualize_weather_forecast(c(48.841319, 2.253076))

}
