% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weather_forecast.R
\name{address_to_gps}
\alias{address_to_gps}
\title{Convertit une adresse en coordonnées GPS.}
\usage{
address_to_gps(adresse)
}
\arguments{
\item{adresse}{Adresse à géocoder}
}
\value{
Coordonnées GPS obtenues à partir de l'adresse
}
\description{
Convertit une adresse en coordonnées GPS.
}
