% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weather_forecast.R
\name{address_to_gps}
\alias{address_to_gps}
\title{Convertir une Adresse en Coordonnées GPS}
\usage{
address_to_gps(location)
}
\arguments{
\item{location}{L'adresse à convertir en coordonnées GPS.}
}
\value{
Un vecteur numérique de taille 2 contenant la latitude et la longitude de l'adresse.
}
\description{
Cette fonction convertit une adresse textuelle en coordonnées GPS à l'aide du service de géocodage OpenStreetMap.
}
\examples{
address_to_gps("Eiffel Tower, Paris, France")
}
