% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address_to_gps.R
\name{address_to_gps}
\alias{address_to_gps}
\title{address_to_gps}
\usage{
address_to_gps(address)
}
\arguments{
\item{address}{caractère de taille 1}
}
\value{
un tibble indiquant la latitude et la longitude de l'adresse indiquée
}
\description{
Cette fonction consiste à convertir une adresse textuelle en coordonnées gps
}
\examples{
#exemple code
address_to_gps("Lorient, France")
}
