% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weather_forecast.R
\name{address_to_gps}
\alias{address_to_gps}
\title{Convertit une adresse en coordonnées GPS.}
\usage{
address_to_gps(adresse)
}
\arguments{
\item{adresse}{Une adresse sous forme de texte.}
}
\value{
Un vecteur numérique de taille 2 avec les coordonnées GPS (latitude, longitude).
}
\description{
Cette fonction prend en entrée une adresse et renvoie les coordonnées GPS sous
forme d'un vecteur numérique de taille 2.
}
