% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weather_forecast.R
\name{address_to_gps}
\alias{address_to_gps}
\title{Convert Address to GPS Coordinates}
\usage{
address_to_gps(adresse, lat, long)
}
\arguments{
\item{adresse}{A character string specifying the address to be converted.}

\item{lat}{numeric Latitude coordinate.}

\item{lon}{numeric Longitude coordinate.}
}
\value{
A numeric vector containing the latitude and longitude coordinates.
}
\description{
This function converts a given address into GPS coordinates using the OpenStreetMap geocoding service.
}
