% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weather_forecast.R
\name{address_to_gps}
\alias{address_to_gps}
\title{Extraction de donnees à partir d'une adresse grace aux fonctions precedentes.}
\usage{
address_to_gps(adresse)
}
\arguments{
\item{adresse}{Vecteur de type character.}
}
\value{
Un tibble contenant les données des previsions meteo, tel que :
- L'heure au tz UTC
- Les donnees de température en Celsuis
- Les donnees de température ressentie en en Celsuis
- La probabilité de pluie en en Pourcentage
- La precipitation en mm
}
\description{
Extraction de donnees à partir d'une adresse grace aux fonctions precedentes.
}
\examples{
#Exemple d'utilisation :

Resultat <- address_to_gps("Champ de Mars, 5 Av. Anatole France, 75007 Paris")
View(Resultat)
}
