% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weather_forecast.R
\name{forecast.character}
\alias{forecast.character}
\title{Verification du type "character" du parametre.}
\usage{
forecast.character(address)
}
\arguments{
\item{address}{Vecteur character de taille 1.}
}
\value{
"L'argument address doit etre de type character et de taille 1." Si le vecteur ne respecte pas le bon typage.
}
\description{
Verification du type "character" du parametre.
}
