% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weather_forecast.R
\name{get_forecast.character}
\alias{get_forecast.character}
\title{Obtient les prévisions météorologiques à partir d'une adresse.}
\usage{
get_forecast.character(adresse)
}
\arguments{
\item{adresse}{Adresse à géocoder}
}
\value{
Les prévisions météorologiques obtenues
}
\description{
Obtient les prévisions météorologiques à partir d'une adresse.
}
