% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weather_forecast.R
\name{get_forecast.character}
\alias{get_forecast.character}
\title{Obtenir des Prévisions Météorologiques pour une Adresse}
\usage{
\method{get_forecast}{character}(location)
}
\arguments{
\item{location}{Une chaîne de caractères représentant une adresse.}
}
\value{
Un tibble contenant les prévisions météorologiques pour l'adresse fournie.
}
\description{
Cette fonction renvoie des prévisions météorologiques pour une adresse textuelle spécifiée.
}
