% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_forecast_character.R
\name{get_forecast.character}
\alias{get_forecast.character}
\title{get_forecast.character}
\usage{
get_forecast.character(address)
}
\arguments{
\item{address}{caractere de taille 1}
}
\value{
un tibble de 5 colonnes et de 168 observations
}
\description{
Obtenir les previsions meteorologiques a partir d'une adresse textuelle.
Faire appel aux fonctions address_to_gps et get_forecast.numeric
}
\examples{
get_forecast.character("Lorient, France")
}
