% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weather_forecast.R
\name{get_forecast.character}
\alias{get_forecast.character}
\title{get_forecast.character}
\usage{
\method{get_forecast}{character}(address)
}
\arguments{
\item{address}{une adresse}
}
\value{
une liste de 2 éléments : une liste de 3 graphiques et une table
}
\description{
Cette fonction permet d'obtenir des prévisions météo facilement lisibles sur
1 semaine, heure par heure, à partir du nom ou de l'adresse d'un lieu. Elles
sont issues de l'API Open Météo.
}
\examples{
get_forecast.character("Parc des Princes")
}
