% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weather_forecast.R
\name{get_forecast.character}
\alias{get_forecast.character}
\title{Application de la verification aux fonctions precedente.}
\usage{
get_forecast.character(address)
}
\arguments{
\item{address}{Vecteur character de taille 1.}
}
\value{
Un tibble contenant les données des previsions meteo, tel que :
- L'heure au tz UTC
- Les donnees de température en Celsuis
- Les donnees de température ressentie en en Celsuis
- La probabilité de pluie en en Pourcentage
- La precipitation en mm
Ou un message d'erreur
}
\description{
Application de la verification aux fonctions precedente.
}
\examples{

#Exemple d'utilisation :
Resultat <- get_forecast.character("Pointe de la Torche, 29120 Plomeur France")
View(Resultat)
#Aucun message d'erreur renvoyé.

get_forecast.character(c(37.7749, -122.4194))
#Attention message d'erreur renvoyé.
}
