% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weather_forecast.R
\name{get_forecast.character}
\alias{get_forecast.character}
\title{get_forecast.character}
\usage{
\method{get_forecast}{character}(address)
}
\arguments{
\item{address}{}
}
\value{
un dataframe
}
\description{
get_forecast.character
}
