% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weather_forecast.R
\name{get_forecast.numeric}
\alias{get_forecast.numeric}
\title{Obtient les prévisions météorologiques à partir de coordonnées GPS.}
\usage{
get_forecast.numeric(xy, ...)
}
\arguments{
\item{xy}{Vecteur numérique avec x les latitudes et y les longitudes}
}
\value{
Un tibble 168x5
}
\description{
Obtient les prévisions météorologiques à partir de coordonnées GPS.
}
