% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weather_forecast.R
\name{get_forecast.numeric}
\alias{get_forecast.numeric}
\title{Obtenir des Prévisions Météorologiques pour des Coordonnées GPS}
\usage{
\method{get_forecast}{numeric}(location)
}
\arguments{
\item{location}{Un vecteur numérique de taille 2 contenant les coordonnées GPS (latitude, longitude).}
}
\value{
Un tibble contenant les prévisions météorologiques pour les coordonnées fournies.
}
\description{
Cette fonction renvoie des prévisions météorologiques pour un ensemble de coordonnées GPS spécifié.
}
