% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weather_forecast.R
\name{get_forecast.numeric}
\alias{get_forecast.numeric}
\title{get_forecast.numeric}
\usage{
\method{get_forecast}{numeric}(xy)
}
\arguments{
\item{xy}{un vecteur comprenant une latitude et une longitude}
}
\value{
une liste de 2 éléments : une liste de 3 graphiques et une table
}
\description{
Cette fonction permet d'obtenir des prévisions météo facilement lisibles sur
1 semaine, heure par heure, à partir des coordonnées gps d'un lieu. Elles
sont issues de l'API Open Météo.
}
\examples{
get_forecast.numeric(c(48.841363,2.253069))

}
