% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weather_forecast.R
\name{get_forecast.numeric}
\alias{get_forecast.numeric}
\title{get_forecast.numeric}
\usage{
\method{get_forecast}{numeric}(xy)
}
\arguments{
\item{xy}{a vector, latitude and longitude}
}
\value{
une liste comprenant les données sous forme de graphiques, puis de tableau de 5 colonnes
}
\description{
Cette fonction permet d'obtenir des prévisions météo facilement lisibles sur
1 semaine, heure par heure, à partir des coordonnées gps d'un lieu. Elles
sont issues de l'API Open Météo.
}
\examples{
get_forecast.numeric(c(48.841363,2.253069))

}
