% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weather_forecast.R
\name{get_forecast.numeric}
\alias{get_forecast.numeric}
\title{Obtient les prévisions météo en fonction des coordonnées GPS.}
\usage{
get_forecast.numeric(xy)
}
\arguments{
\item{xy}{Un vecteur numérique de taille 2 représentant les coordonnées GPS.}
}
\value{
Un tibble avec les prévisions météo.
}
\description{
Cette fonction prend en entrée un vecteur numérique de taille 2 (latitude, longitude)
et renvoie les prévisions météo sous forme de tibble.
}
\seealso{
\code{\link{perform_request}}, \code{\link{unnest_response}}
}
