% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weather_forecast.R
\name{get_forecast.numeric}
\alias{get_forecast.numeric}
\title{get_forecast.numeric}
\usage{
\method{get_forecast}{numeric}(coordinates)
}
\arguments{
\item{coordinates}{}
}
\value{
un dataframe
}
\description{
get_forecast.numeric
}
