% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weather_forecast.R
\name{get_forecast}
\alias{get_forecast}
\title{get_forecast}
\usage{
get_forecast(x)
}
\arguments{
\item{location}{Emplacement pour obtenir les prévisions (Adresse, nom de site olympique ou coordonnées GPS).
\itemize{
\item Pour get_forecast.character : un caractère de taille 1.
\item Pour get_forecast.numeric : un vecteur numérique de taille 2 (latitude, longitude).
}}
}
\value{
Un tibble contenant les prévisions météorologiques.
}
\description{
Cette fonction générique permet d'obtenir des prévisions météorologiques en fonction d'une localisation.
Elle a deux implémentations spécifiques :
\itemize{
\item get_forecast.character : pour obtenir des prévisions à partir d'un nom de site olympique ou d'une adresse.
\item get_forecast.numeric : pour obtenir des prévisions à partir des coordonnées GPS.
}
}
\examples{
Exemples
\dontrun{
  # Exemples d'utilisation
  get_forecast("Nantes")
  get_forecast(c(48.85, 2.35))
}

}
\seealso{
\code{\link{get_forecast.character}}, \code{\link{get_forecast.numeric}},
\code{\link{perform_request}}, \code{\link{unnest_response}}
}
