% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_forecast.R
\name{get_forecast}
\alias{get_forecast}
\title{Données météo}
\usage{
get_forecast(x, y = NULL)
}
\arguments{
\item{x}{Latitude (numérique) ou adresse (caractère).}

\item{y}{Longitude (numérique). Obligatoire uniquement si x est la latitude.}
}
\value{
Une tibble contenant des données de prévisions météorologiques.
}
\description{
Cette fonction permet d'obtenir des données de prévisions météorologiques en fonction de la latitude et de la longitude ou d'une adresse.
}
\examples{
get_forecast(42.35, 2.65)
get_forecast("Paris")
}
