% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weather_forecast.R
\name{get_forecast}
\alias{get_forecast}
\title{Title: A function for making weather forecasts}
\usage{
get_forecast(value)
}
\arguments{
\item{value}{A string of length 1 or a numeric vector of length 2 for Address or Latitude and Longitude values.}
}
\value{
A tibble containing the forecast about the temperature, the apparent temperature, the precipitation probability, and the precipitation in mm.
}
\description{
This function uses the value of the object provided to the function and returns weather forecasts.
}
\examples{
get_forecast("1600 Amphitheatre Parkway, Mountain View, CA")
get_forecast(c(68, 4.2))


}
\concept{weather functions}
