% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weather_forecast.R
\name{get_forecast}
\alias{get_forecast}
\title{Obtenir des Prévisions Météorologiques}
\usage{
get_forecast(location)
}
\arguments{
\item{location}{Soit un vecteur numérique de taille 2 contenant les coordonnées GPS (latitude, longitude),
soit une chaîne de caractères représentant une adresse.}
}
\value{
Un tibble contenant les prévisions météorologiques pour la localisation spécifiée.
}
\description{
Cette fonction générique renvoie des prévisions météorologiques basées sur la localisation fournie.
La localisation peut être fournie sous forme de coordonnées GPS (vecteur numérique) ou d'une adresse (chaîne de caractères).
}
\examples{
get_forecast(c(48.85, 2.35)) # Coordonnées GPS de Paris
get_forecast("Eiffel Tower, Paris, France") # Adresse textuelle
}
