% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weather_forecast.R
\name{get_forecast}
\alias{get_forecast}
\title{data meteo}
\usage{
get_forecast(x, y = NULL)
}
\arguments{
\item{x}{Latitude (num) ou adresse (string)}

\item{y}{Longitude (num) (Obligatoire uniquement si x est la latitude).}
}
\value{
Un tibble contenant des donnees de previsions meteorologiques
}
\description{
permet d obtenir des donnes de previsions meteorologiques en fonction de la lat. et de la long. ou d une adresse
}
\examples{
get_forecast(26.80, 4.83)
get_forecast("nantes")
}
