% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weather_forecast.R
\name{get_forecast}
\alias{get_forecast}
\title{Obtient les prévisions météo en fonction des coordonnées GPS ou de l'adresse.}
\usage{
get_forecast(x)
}
\arguments{
\item{x}{Un vecteur numérique de taille 2 représentant les coordonnées GPS (latitude, longitude) ou une adresse spécifiée en tant que caractère.}

\item{address}{Une adresse spécifiée en tant que caractère.}
}
\value{
Une tibble contenant les prévisions météo, comprenant la date, l'heure UTC, la température, la
température ressentie, la probabilité de précipitation, et la quantité de précipitation.

Une tibble contenant les prévisions météo.
}
\description{
Cette fonction est générique et permet d'obtenir les prévisions météo en fonction des coordonnées GPS
(latitude, longitude) ou d'une adresse spécifiée.
}
\examples{
# Obtenir les prévisions météo pour des coordonnées GPS
xy_coordinates <- c(48.85, 2.35)
forecast_result <- get_forecast(xy_coordinates)
print(forecast_result)

# Obtenir les prévisions météo pour une adresse
address_result <- get_forecast("9 Quai Henri Barbusse, Nantes, 44000, FRANCE")
print(address_result)

}
\seealso{
\code{\link{get_forecast.numeric}} et \code{\link{get_forecast.character}}
}
