% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weather_forecast.R
\name{get_forecast}
\alias{get_forecast}
\title{Obtenir les previsions meteo à partir d'une adresse.}
\usage{
get_forecast(x)
}
\arguments{
\item{x}{Adresse (character) ou coordonnées (numeric).}
}
\value{
Un tibble contenant les données des previsions meteo, tel que :
- L'heure au tz UTC
- Les donnees de température en Celsuis
- Les donnees de température ressentie en en Celsuis
- La probabilité de pluie en en Pourcentage
- La precipitation en mm
}
\description{
Obtenir les previsions meteo à partir d'une adresse.
}
\examples{
# Exemple d'utilisation :
      - resultats_finals <- get_forecast("15 rue louis Gaudin, Nantes,FRANCE")
        print(resultats_finals)

      - resultats_finals2 <- get_forecast(c(37.7749, -122.4194))
        print(resultats_finals2)
}
