% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weather_forecast.R
\name{get_forecast}
\alias{get_forecast}
\title{Prévisions Météo}
\usage{
get_forecast(add_ou_coord)
}
\arguments{
\item{add_ou_coord}{Adresse littérale ou coordonnées géographiques du lieu souhaité}
}
\value{
un tibble contenant la météo horaire de l'endroit en question des 7 prochains jours
}
\description{
Cette fonction permet d'obtenir les prévisions météo en temps réel d'un lieu pour chaque heure des 7 prochains jours. Voici les 4 renseignements obtenus : la "temperature_celsius", la "temperature_ressentie_celsius", la "precipation_proba" et les "precipitation".
}
\examples{
get_forecast("Nantes")

coord = c(47.218371, -1.553621)
get_forecast(coord)

}
