% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weather_forecast.R
\name{get_forecast}
\alias{get_forecast}
\title{get_forecast}
\usage{
get_forecast(location)
}
\arguments{
\item{location}{}
}
\value{
un dataframe
}
\description{
get_forecast
}
\examples{
get_forecast("Paris")
}
