% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weather_forecast.R
\name{get_gps_coordinate}
\alias{get_gps_coordinate}
\title{Obtient les coordonnées GPS à partir d'une adresse.}
\usage{
get_gps_coordinate(adresse)
}
\arguments{
\item{adresse}{Adresse à géocoder}
}
\value{
Coordonnées GPS obtenues à partir de l'adresse
}
\description{
Obtient les coordonnées GPS à partir d'une adresse.
}
