% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weather_forecast.R
\name{get_gps_coordinate}
\alias{get_gps_coordinate}
\title{Obtenir les Coordonnées GPS pour une Adresse}
\usage{
get_gps_coordinate(location)
}
\arguments{
\item{location}{L'adresse pour laquelle les coordonnées GPS sont demandées.}
}
\value{
Un vecteur numérique contenant la latitude et la longitude de l'adresse.
}
\description{
Fonction enveloppe appelant \code{address_to_gps} pour obtenir les coordonnées GPS d'une adresse donnée.
}
\examples{
get_gps_coordinate("Eiffel Tower, Paris, France")
}
