% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gps_coordinate.R
\name{get_gps_coordinate}
\alias{get_gps_coordinate}
\title{get_gps_coordinate}
\usage{
get_gps_coordinate(address)
}
\arguments{
\item{address}{caractere de taille 1}
}
\value{
un vecteur avec les coordonnées x et y.
}
\description{
Cette fonction consiste egalement a indiquer les coordonnees GPS d'une adresse.
Elle est interne car elle fait appel a une autre fonction (address_to_gps).
x renvoie la longitude et y la latitude
}
\examples{
get_gps_coordinate("Lorient, France")
}
