% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weather_forecast.R
\name{get_gps_coordinate}
\alias{get_gps_coordinate}
\title{get_gps_coordinate}
\usage{
get_gps_coordinate(adresse)
}
\arguments{
\item{adresse}{une adresse}
}
\value{
un vecteur de coordonnées (lat, long)
}
\description{
Cette fonction permet d'obtenir un vecteur de coordonnées (latitude et
longitude) à partir d'une adresse ou d'un nom de lieu.
}
\examples{
get_gps_coordinate("Parc des Princes")

}
