% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weather_forecast.R
\name{get_gps_coordinate}
\alias{get_gps_coordinate}
\title{Renvoie les coordonnees x et y associees à une adresse donnee.}
\usage{
get_gps_coordinate(adresse)
}
\arguments{
\item{adresse}{Vecteur de type character.}
}
\value{
Les coordonnees x et y associees a une adresse donnee.
}
\description{
Renvoie les coordonnees x et y associees à une adresse donnee.
}
\examples{
#Exemple d'utilisation :

 Resultat <- get_gps_coordinate("Champ de Mars, 5 Av. Anatole France, 75007 Paris")
 View(Resultat)
}
