% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weather_forecast.R
\name{get_graphs}
\alias{get_graphs}
\title{get_graphs}
\usage{
get_graphs(data)
}
\arguments{
\item{data}{un tibble avec 5 colonnes}
}
\value{
une liste de 3 graphiques
}
\description{
Cette fonction affiche 3 graphiques représentant chacun l'évolution des
caractéristiques météo prévues.
}
\examples{
get_graphs(unnest_response(perform_request(48.85,2.35)))

}
