% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weather_forecast.R
\name{get_weather_forecast}
\alias{get_weather_forecast}
\title{Prévision météo des sites olympiques}
\usage{
get_weather_forecast(localisation)
}
\arguments{
\item{localisation}{Latitude et longitude de la coordonnée GPS (vecteur numérique) ou adresse (chaîne de caractères).}
}
\value{
Un tibble contenant les prévisions météo.
}
\description{
Cette fonction permet de récupérer une table de prévisions météo pour une coordonnée GPS donnée. Ces prévisions météo incluent des informations telles que la température, la probabilité de précipitations, etc.
}
\examples{

get_weather_forecast(c(48.85, 2.35))
get_weather_forecast("Paris")

}
