% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_function.R
\name{graph_function}
\alias{graph_function}
\title{Génère un graphique basé sur les données fournies.}
\usage{
graph_function(unnested_table)
}
\arguments{
\item{unnested_table}{Un tibble avec les données nécessaires pour le graphique.}
}
\value{
Un objet ggplot.
}
\description{
Génère un graphique basé sur les données fournies.
}
