% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weather_forecast.R
\name{perform_request}
\alias{perform_request}
\title{Requête pour récupérer les données de prévisions météorologiques}
\usage{
perform_request(latitude, longitude)
}
\arguments{
\item{latitude}{Latitude (numérique)}

\item{longitude}{Longitude (numérique)}
}
\value{
Un tibble contenant les données de prévisions météorologiques
}
\description{
Requête pour récupérer les données de prévisions météorologiques
}
