% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_request.R
\name{perform_request}
\alias{perform_request}
\title{Fonction interne pour effectuer la requête API et obtenir les données brutes}
\usage{
perform_request(lat,lon)
}
\arguments{
\item{lat}{Latitude de la coordonnée GPS.}

\item{lon}{Longitude de la coordonnée GPS.}
}
\value{
Les données brutes de prévisions météo.
}
\description{
Cette fonction interne effectue une requête API pour récupérer les données brutes de prévisions météo.
}
\examples{
perform_request(48.841319, 2.253076)

}
