% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weather_forecast.R
\name{perform_request}
\alias{perform_request}
\title{Obtenir les prévisions météorologiques actualisées à partir de l'API.
Cette fonction récupère les prévisions météorologiques les plus récentes en utilisant les coordonnées GPS (latitude, longitude) spécifiées en entrée.}
\usage{
perform_request(lat, lon)
}
\arguments{
\item{latitude}{La latitude (numeric).}

\item{longitude}{La longitude (numeric).}
}
\value{
Un tibble contenant les prévisions météorologiques.
}
\description{
Obtenir les prévisions météorologiques actualisées à partir de l'API.
Cette fonction récupère les prévisions météorologiques les plus récentes en utilisant les coordonnées GPS (latitude, longitude) spécifiées en entrée.
}
