% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weather_forecast.R
\name{perform_request}
\alias{perform_request}
\title{Effectuer une Requête à l'API Météo}
\usage{
perform_request(latitude, longitude)
}
\arguments{
\item{latitude}{La latitude de l'emplacement pour lequel obtenir les prévisions.}

\item{longitude}{La longitude de l'emplacement pour lequel obtenir les prévisions.}
}
\value{
Un tibble contenant les données météorologiques horaires pour l'emplacement spécifié.
}
\description{
Cette fonction envoie une requête à l'API Open-Meteo en utilisant des coordonnées GPS spécifiques
et retourne des données météorologiques horaires sous forme de tibble.
}
\examples{

perform_request(latitude = 48.85, longitude = 2.35)
}
