% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weather_forecast.R
\name{perform_request}
\alias{perform_request}
\title{perform_request}
\usage{
perform_request(latitude, longitude)
}
\arguments{
\item{latitude}{latitude (numeric)}

\item{longitude}{longitude (numeric)}
}
\value{
un tibble avec 5 lignes
}
\description{
Cette fonction permet d'effectuer une requête sur l'API Open Meteo à partir
de coordonnées gps. Elle porte sur l'heure, la température, la température
ressentie, la probabilité qu'il pleuve et la quantité de pluie prévue sur 1
semaine.
}
\examples{
perform_request(48.85,2.35)

}
