% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weather_forecast.R
\name{perform_request}
\alias{perform_request}
\title{Effectue une requête à l'API pour obtenir les prévisions météo.}
\usage{
perform_request(lat, lon)
}
\arguments{
\item{lat}{La latitude (numeric).}

\item{lon}{La longitude (numeric).}
}
\description{
Cette fonction prend en entrée les coordonnées GPS (latitude, longitude) et retourne
les prévisions météo sous forme de tibble.
}
