% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weather_forecast.R
\name{perform_request}
\alias{perform_request}
\title{perform_request}
\usage{
perform_request(lat, lon)
}
\arguments{
\item{lat}{numeric Latitude coordinate.}

\item{lon}{numeric Longitude coordinate.}
}
\value{
A tibble with weather forecast information.
}
\description{
This function is designed to create a tibble by making a request to an API.
}
\details{
This function takes latitude and longitude coordinates as input, constructs a URL to query a weather forecast API, and retrieves weather forecast data in the form of a tibble.
}
\examples{
perform_request(48.85, 2.35)

}
