% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weather_forecast.R
\name{perform_request}
\alias{perform_request}
\title{Extraction de donnees meteorologiques d'un point donne grace à ses coordonnees.}
\usage{
perform_request(latitude, longitude)
}
\arguments{
\item{latitude}{coordonnees gps de la latitude}

\item{longitude}{coordonnees gps de la longitude}
}
\description{
Extraction de donnees meteorologiques d'un point donne grace à ses coordonnees.
}
\examples{
#Exemple d'utilisation :

Resultat <- perform_request(37.7749, -122.4194)
View(Resultat)
}
