% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weather_forecast.R
\name{unnest_response}
\alias{unnest_response}
\title{Décompacte les données de réponse des prévisions météorologiques.}
\usage{
unnest_response(resp)
}
\arguments{
\item{resp}{Résultat de la fonction précédente, tibble 5x9}
}
\value{
Un nouveau tibble comprenant 168 observations et 5 variables
}
\description{
Décompacte les données de réponse des prévisions météorologiques.
}
