% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unnest_response.R
\name{unnest_response}
\alias{unnest_response}
\title{Fonction interne pour transformer les données brutes en format de prévisions météo}
\usage{
unnest_response(donnees_brute)
}
\arguments{
\item{donnees_brute}{Les données de prévisions météo.}
}
\value{
Un tibble contenant les prévisions météo formatées.
}
\description{
Cette fonction interne prend les données brutes de prévisions météo et les transforme en un format plus lisible.
}
\examples{
unnest_response(olympicsWeather::perform_request(48.841319, 2.253076))

}
