% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weather_forecast.R
\name{unnest_response}
\alias{unnest_response}
\title{Remodeler la tibble obtenue de la requête pour correspondre à une structure donnée.Cette fonction prend en entrée la tibble résultant de la requête et la transforme pour générer une nouvelle tibble contenant les données requises.}
\usage{
unnest_response(extraction)
}
\arguments{
\item{extraction}{La méthode d'extraction à utiliser pour extraire les différentes colonnes de la liste de données.}
}
\value{
Une tibble remodelée selon la structure spécifiée.
}
\description{
Remodeler la tibble obtenue de la requête pour correspondre à une structure donnée.Cette fonction prend en entrée la tibble résultant de la requête et la transforme pour générer une nouvelle tibble contenant les données requises.
}
