% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unnest_response.R
\name{unnest_response}
\alias{unnest_response}
\title{unnest_response}
\usage{
unnest_response(resp)
}
\arguments{
\item{resp}{a tibble}
}
\value{
Un tableau des 5 listes de la colonne hourly
}
\description{
Cette fonction permet de retourner un tableau ayant les listes de la colonne hourly en colonnes avec comme entrée les coordonnées de la fonction précédente
}
\examples{
resp <- perform_request(48.85,2.35)
resp |>  unnest_response() -> y
plot(y$temperature_celsius, type="l")
}
