% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weather_forecast.R
\name{unnest_response}
\alias{unnest_response}
\title{unnest_response}
\usage{
unnest_response(resp)
}
\arguments{
\item{resp}{un tibble avec 5 lignes}
}
\value{
un tibble avec 5 colonnes
}
\description{
Cette fonction permet d'améliorer la lisibilité des résultats obtenus grâce à
la fonction perform_request. Elle permet de créer un tibble de 5 colonnes où
chacune d'elles est une catégorie d'informations et chaque ligne correspond à
une prévision, donc à 1 heure.
}
\examples{
unnest_response(perform_request(48.85,2.35))

}
