% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weather_forecast.R
\name{unnest_response}
\alias{unnest_response}
\title{unnest_response}
\usage{
unnest_response(resp)
}
\arguments{
\item{resp}{a tibble with 5 rows, from the request on the API}
}
\value{
a tibble with 5 columns, more readable
}
\description{
Cette fonction permet d'améliorer la lisibilité des réultats obtenus grâce à
la fonction perform_request. Elle permet de créer un tible de 5 colonnes ou
chacune d'elles est une catégorie d'information et chaque ligne correspond à
une prévision et donc à 1 heure.
}
\examples{
unnest_response(perform_request(48.85,2.35))

}
