% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weather_forecast.R
\name{unnest_response}
\alias{unnest_response}
\title{Transforme la tibble obtenue de la requête en une structure de données spécifique.}
\usage{
unnest_response(extraction)
}
\arguments{
\item{extraction}{sert a extraire les différentes colonnes de la liste de données}
}
\description{
Cette fonction prend en entrée la tibble obtenue de la requête et la transforme pour
obtenir une nouvelle tibble avec les informations nécessaires.
}
