% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weather_forecast.R
\name{unnest_response}
\alias{unnest_response}
\title{unnest_response}
\usage{
unnest_response(resp)
}
\arguments{
\item{resp}{Weather API response, typically in the form of a list.}
}
\value{
A tibble with columns for date and time, temperature in Celsius, apparent temperature in Celsius, precipitation probability, and precipitation amount.
}
\description{
This function is designed to flatten weather API response data into a tibble.
}
\details{
This function takes as input the weather API response, typically in the form of a list, and extracts relevant data to flatten into a tibble with the following columns:
  - date_time: Date and time of the weather forecast.
  - temperature_celsius: Temperature in Celsius.
  - apparent_temperature_celsius: Apparent temperature in Celsius.
  - precipation_probability: Probability of precipitation.
  - precipitation: Amount of precipitation.

The function then returns this flattened tibble with the data.
}
\examples{
response <- list(
  time = c("2024-05-23T00:00:00Z", "2024-05-23T01:00:00Z"),
  temperature_2m = c(15.0, 14.5),
  apparent_temperature = c(15.0, 14.0),
  precipitation_probability = c(0.1, 0.2),
  precipitation = c(0.0, 0.1)
)
unnest_response(response)

}
\seealso{
[as_tibble()]
}
