% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weather_forecast.R
\name{unnest_response}
\alias{unnest_response}
\title{Formatage des donnees recueillis grace à la fonction perform_request}
\usage{
unnest_response(Extraction)
}
\arguments{
\item{Extraction}{Vecteur de coordonnees de type (latitude, longitude) apres application de la fonction perform_request.}
}
\value{
Un tibble contenant les données des previsions meteo, tel que :
- L'heure au tz UTC
- Les donnees de température en Celsuis
- Les donnees de température ressentie en en Celsuis
- La probabilité de pluie en en Pourcentage
- La precipitation en mm
}
\description{
Formatage des donnees recueillis grace à la fonction perform_request
}
\examples{
#Exemple d'utilisation :

Extraction <- perform_request(37.7749, -122.4194)
Resultat <- unnest_response(Extraction)
View(Resultat)
}
